requirements.txt
aiofiles==0.6.0
attrs==20.3.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
ecdsa==0.14.1
fastapi==0.63.0
h11==0.12.0
idna==2.10
importlib-metadata==3.7.0
iniconfig==1.1.1
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
passlib==1.7.4
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
pyparsing==2.4.7
pytest==6.2.2
python-jose==3.2.0
python-multipart==0.0.5
requests==2.25.1
rsa==4.7.1
six==1.15.0
SQLAlchemy==1.3.23
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.3
zipp==3.4.1
